+++ mc/1-8---1-12/src/main/java/dev/redstudio/redcore/asm/RedLoadingPlugin.java
@@ -1,0 +1,36 @@
+package dev.redstudio.redcore.asm;
+
+import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin;
+
+import java.util.Map;
+
+/// A base class for [IFMLLoadingPlugin] implementing standard stubs.
+///
+/// @author Luna Mira Lage (Desoroxxx)
+/// @since 0.7
+public abstract class RedLoadingPlugin implements IFMLLoadingPlugin {
+
+	@Override
+	public String[] getASMTransformerClass() {
+		return null;
+	}
+
+	@Override
+	public String getModContainerClass() {
+		return null;
+	}
+
+	@Override
+	public String getSetupClass() {
+		return null;
+	}
+
+	@Override
+	public void injectData(final Map<String, Object> data) {
+	}
+
+	@Override
+	public String getAccessTransformerClass() {
+		return null;
+	}
+}